only get env file if not there

+++ scripts/setup.sh
@@ -37,11 +37,13 @@ checkIfAbleToTalkToAWS() {
 }
 
 fetchEnv() {
-  if [[ ${STAGE} != "PROD" ]]; then
-      aws s3 cp s3://editorial-tools-integration-tests-dist/env.dev.json ${DIR}/../env.json --profile media-service
-  else
-      aws s3 cp s3://editorial-tools-integration-tests-dist/env.dev.json ${DIR}/../env.json
-  fi
+    if [[ ! -f ${DIR}/../env.json ]]; then
+        if [[ ${STAGE} != "PROD"  ]]; then
+            aws s3 cp s3://editorial-tools-integration-tests-dist/env.dev.json ${DIR}/../env.json --profile media-service
+        else
+            aws s3 cp s3://editorial-tools-integration-tests-dist/env.dev.json ${DIR}/../env.json
+        fi
+    fi
 }
 
 checkForNodeModules
